Allow repeated searching for cooldown text with OmniCC
Previously if no font string was found on an aura button's cooldown, no further searches were made. Now, if OmniCC is detected, a search will be performed on every button update until the font string is found. This resolves an issue wherein OmniCC seemed to delay creation of the text on the first button created, resulting in that button's text not getting styled by oUF_Phanx.